Introduce Rambdax
I'd gone back and forth a bit on what utility library I would want to
use. I like Ramda, but it's implementations are slower compared to
other options and it doesn't tree-shake well.

I found Rambda which seemed nice, and was considering Lodash just
because it's so well known. But Rambdax seemed like the perfect
fit:
  - Small in size
  - Good performance
  - Extensive API