Setup tests with Rollup
Adds the ability to run testing with Rollup. There's a trick to this
because Rollup is a lot more specific about how it operates than
Webpack is. Tests work in the following way:
  - `tsc` is used to transpile all Typescript files into JavaScript in
    a "shadow" directory of src/ called build/
  - Karma looks for JavaScript files in build/ and runs it's own
    Rollup config over them
  - the `test:watch` command runs tsc and karma concurrently in watch
    mode

This is plenty fast right now, but an eye should be kept on it in case
it starts performing slowly.