Add test scripts for CI
This completes work on getting testing setup for the project.

Close #1
Close #2

Rollup has been sufficient for the project, so I don't see any reason
to switch right now.

Close #5

Also this partially addresses #4, though for it to be fully complete
the project needs it's own Docker image for a build environment.