Don't use PhantomJS in CI environments
The project uses Alpine Linux for it's CI environment. This is great
for a lot of reasons but it comes with not so much downsides, but a
double edged sword. Alpine uses the musl C library; it's a
significantly lighter and faster library, and those benefits carry
down to binaries built on it. But not everything is compatible with
it, PhantomJS being one of those.

Right now there's no reason to use Karma in a CI environment; it's
more of a local development tool. In a CI environment, tests will just
be run with Mocha.

Close #4