Create base code for Character Attributes
In order to accomplish this a collection of other changes had to make
it in as well. These include:
  - Adding a shim to Karma for PhantomJS
  - Adding two new scripts for running tests
  - Removed the "test Module" that was created just to know that
    testing was setup correctly.
  - Add libs to tsconfig so that newer language features could be used
  - Stop enforcing "no magic numbers", it was getting out of hand

Magic numbers are usually horribly annoying, but in this case, where
there are so many arbitrary numbers, most of a function body would
have just been names for numbers. It would be good to turn it back on
but right now it's not a priority.