Create basic code for skills
Compared to other recent work this is weaker work. I have ideas for
what to do about improving it; but as it stands it's rough.
Serviceable, but rough. When I get around to a code clean-up after
getting all this early work done these files are going to need some
extra attention.