Improve Karma config, rework dependencies
This commit is unfortunately big, but it is all conceptually related
work.

1. Improve Karma's configuration.
This hadn't been getting much love for a while. And the initial
configuration that I wrote was starting to break down, and become
flakier as code was added to the project. One of the main benefits
Karma provides is the ability to ensure the same code is working in
multiple browsers at once. Introducing Rambdax into the test suite
caused a breakage with legacy browsers.

The problem is that Rambdax doesn't provide ES5 compliant code in the
dist folder, and bundlers don't compile imported node module code. So
ES6 code was getting inserted into the final bundle. I tried to work
around it for a while, but everything I came up with was too
hack-y. That lead to..

2. Reworking/trimming of dependencies
While Ramda is larger slightly larger than Rambdax, they're both
tree shakable, and the final build size shouldn't be noticeably
different. Ramda is also a bit less optimized in most areas, but not
to a level where it's likely to be a bottle neck. Importantly though,
Ramda distributes standard, ES5 compliant code.

Further trimming was possible considering PhantomJS isn't usable
within the CI environment, so it didn't make sense to keep it as a
dependency. And without the issues of Rambdax in the bundle Bublé
stopped being necessary as well. The single Lodash function was very
easily home rolled, allowing for that to be dropped as well.

3. Adding some more code to the index file
The build check had a blind spot, in the sense that none of the new
MekWarrior4 code was being built along with the index file. It still
built fine, but after spending time working with Karma and how it's
configuration was allowed to slip it made sense to put something into
the index file to ensure the build is actually working as intended.