Load shim file in legacy browsers
The "nomodule" attribute will be ignored in any modern browser, but in
legacy browsers a corejs shim will get loaded in. The fact that it's
being loaded in from node_modules will need to change in the future
when it goes to a production environment, but this does the job for
now.