Class based refactor
The original idea with this code was to write it in a more functional
way, with just interfaces that data could be typed as, and then
writing a bunch of functions to operate on that data. This worked
pretty well at first, but as I started getting into some of the
larger data structures I wasn't liking how the code was turning out,
many of the functions were beginning to need a lot of parameters
making them feel clumsy to use.

Instead I've gone back and rewritten all the old interfaces to be
classes instead, and to work in a more object oriented way. I think
there's still room for improvement with the specs like using
`beforeEach` blocks; but they are effectively exercising the code so
they'll do for now.