Initial version of Names DB
This is a very basic version of Names DB. It can connect to a database
with environment variables, query a view, and turn the results into
JSON; MVP kind of stuff.

There are a few rough parts I'll work out later.

The first issue is that two or three of the dependencies in the toml
file are unnecessary, and can probably be stripped out later on.

The second is that everything is in a single file right now, which
will need to be broken apart going forward. however this is very much
a micro-service, so I'm thinking it's probably looking to be 2-5
source files total.

Third and final issue is that there are no tests. I'm thinking that's
a later addition because the surface area of this is so small it's
extremely easy to manually test it. So I probably won't add testing
until I'm ready for the CI to start verifying it for deploys and
such.