Add a lot of common files to the .gitignore
This is preemptive, but I wanted to ensure that a bunch of weird files
don't end up in future merge requests, so I've gone through and added
a bunch of commonly created files by editors and operating systems.

The list can't be exhaustive or I'd have thousands of lines in the
ignore file; but I think this covers the most commonly used editors.