Stop testing the ability to build with a release build
Initially I set the build to use a release build without thinking
deeply about it. The point of the build step is to ensure that any
language involved in MekManager that has a compile step will build.

When I made the choice my mind was on my day-to-day work which
involves Angular 6, where it's default compilation and "optimized"
compilation aren't the same thing. A project that compiles 100% okay
by default, may not compile when "optimized".

None of that's the case with Rust. In this case I was just wasting
time on a longer, unnecessary optimized build when I only intended to
prove that the code compiles.