Initial work for life module system
This is the basic idea I'm going for with it, there's still some work
to be done figuring out how to handle all of a characters life modules
and all the specific rules that might come with it. Specifically how
to ensure each of them get validated. But this is a strong start.