Update Character Creation to use Validator system
I whiteboard'd this out the other day, and I'm much happier with how
this works than what I originally had going on in the character
creation harness.

I've been working hard to keep everything as generic as possible but
some of this is too specific to be stored all as data. What now exists
is a set of "validators" for the rules of the game. They are a bit
expensive because they're pretty loopy, but the collections they loop
over are always going to be pretty small, usually less than 10 items,
so I'm not concerned about it. I'll start getting concerned if I
notice bad performance in the future.

While this isn't what I'd consider ideal (being able to add new
functionality through data alone), it does make it very clear *how* to
add new functionality. If a new rule is an exception, than the "base
validator" needs to be aware of the exception. Otherwise, you make a
"special validator" for the rule.