Remove reliance on Ramda in MekWarrior 4 code
I have an idea that at some point in the future the MekWarrior4
directory, and maybe others will become a library, and the stuff in
the root of the src/ directory will go on being it's own thing.

If that's going to be the case I'd like to keep that library as light
as possible.