Update rule structure
This adds a new validator for asserting an affiliation cannot be a
character's sole affiliation.

The bigger change here is how rules work. Rules are no longer just a
part of an enum, but are a class of their own. That enum still exists
as the name of the rule. But other rules I'm going to start adding
will need configuration values, hence the addition.