Stop enforcing capitalized comment lines
This is less that I stopped agreeing with the spirit of this and more
that the way tslint enforces it doesn't match how I envision it. What
happens is that if there are lot of single line comments next to
each other tslint will complain if each of them don't start with a
capital letter, even though they're all part of one paragraph.