Add last two special Clan rules
This adds the last two rule validators out of the core rules; at least
that I was able to identify initially.

Both of these last two validators are special cases for the Clans.

The first is in cases where a Clan must use a specific phenotype
over another.

The second is a validator that forces Clan warriors to have taken a
specific trait.