First pass at creating the core library
This take most of the work that I've been doing in the web directory and
moves it out to a new directory specifically for the core library.

There's still some things I need to figure out like what the best way to
go about packaging this is. I was taking inspiration from some bigger
projects but the people who work on those are cover a lot more cases
than I'd likely need to.

I think I'll need a couple more commits to get this right, but this is a
good MVP.