Refactor Attributes
The longer I was working with the original version of Attributes that
I'd created two things became clear:
  1. It behaved like no other part of the system, and felt completely
     out of place to interact with it.
  2. The abstraction was very leaky/bad, and it felt clumsy to work
     with.

This refactors Attributes to be a class like the rest of the system.
There's more room for improvement here. Because of how the old system
worked the Character class calls out to set XP directly, rather than
adding or removing from it.

I could have done that in this commit but I wanted to retain the
behavior of the original as much as possible here.