Add interfaces
I was beginning to notice that I was doing several things that were very
similar, but I wasn't being 100% consistent about how I did it. Or
writing it in a way that allowed me to be more generic about how I
expressed the idea. Interfaces are a good way to handle this, especially
as how this whole subsystem works is starting to come into focus.